+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Breaking Changes
 
 * Let pipeline perform raw data parsing and indices calculation on combined raw data from single tasks ([#73](https://github.com/psychelzh/tarflow.iquizoo/issues/73)).
+* Let `prepare_fetch_data()` signal error when `contents` contains no data.
 
 # tarflow.iquizoo 3.4.0